# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalogs API Documentation",
        "description": "This is the API documentation for my Laravel application using Swagger/OpenAPI 3.0.",
        "contact": {
            "name": "API Support Team",
            "email": "ptryshaf@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/author": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get list of authors",
                "operationId": "96fd7fa3ca2215afa7bc306907175ac9",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search keyword for author name or bio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of authors retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List authors successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "bio": {
                                                        "type": "string",
                                                        "example": "Lorem ipsum dolor sit amet"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-01-01T00:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-01-01T00:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Create a new author",
                "operationId": "5d7a5ee4c26df8bab9e8bed5f32ac0d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "bio"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Author created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/author/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get detail of a author by ID",
                "operationId": "f5da13c0fe7f7b4671318c86ce983ab9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Author retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "bio": {
                                            "type": "string",
                                            "example": "Lorem ipsum dolor sit amet"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-01-01T00:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-01-01T00:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "author not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Update author",
                "operationId": "78fb2f56bc5bb5af527c98839faebe7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "bio"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "author updated successfully"
                    },
                    "404": {
                        "description": "author not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Delete a author by ID",
                "operationId": "2f2f01c9de9ff82ede9ba2f3c36b25a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "author ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "author deleted successfully"
                    },
                    "404": {
                        "description": "author not found"
                    }
                }
            }
        },
...
    },
...
}